Enhance dashboard functionality with Backup Tools integration
- Added `BackupToolsURL` parameter to the `NewService` and `NewAPI` constructors for improved integration with Backup Tools.
- Updated `GetDashboardStats` method to enrich AutoSync card data using Backup Tools statistics.
- Refactored `Auth` controller to streamline dashboard data retrieval and response handling.
- Removed redundant parameters and methods related to Backup Tools from the `Auth` controller, simplifying the codebase.
